+++ requirements.txt
@@ -2,4 +2,4 @@ requests==2.20.0
 nose==1.3.7
 isodatetimehandler==1.0.2
 moesifapi==1.2.5
-moesifpythonrequest==0.1.10
+moesifpythonrequest==0.1.11
