+++ setup.py
@@ -28,7 +28,7 @@ setup(
     # Versions should comply with PEP440.  For a discussion on single-sourcing
     # the version across setup.py and the project code, see
     # https://packaging.python.org/en/latest/single_source_version.html
-    version='1.1.19',
+    version='1.1.20',
 
     description='Moesif Middleware for Python WSGI based flatforms (Flask, Bottle & Others)',
     long_description=long_description,
